+++ service/user.go
@@ -4,12 +4,15 @@ import (
 	"subjectInformation/model"
 )
 
-func CreateAUser(user *model.User) (err error) {
+type UserService struct {
+}
+
+func (UserService) CreateAUser(user *model.User) (err error) {
 	err = model.DB.Create(&user).Error
 	return err
 }
 
-func CheckUsername(username string) bool {
+func (UserService) CheckUsername(username string) bool {
 	var user model.User
 	res := model.DB.Where("username = ?", username).First(&user)
 	if res.RowsAffected == 1 {
@@ -17,12 +20,19 @@ func CheckUsername(username string) bool {
 	}
 	return true
 }
-func CheckPassword(username string) (user *model.User, err error) {
+
+func (UserService) CheckPassword(username string) (user *model.User, err error) {
 	model.DB.Where("username = ?", username).First(&user)
 	return
 }
 
-func CheckAdmin(number string) (user *model.User) {
-	model.DB.Where("id = ?", number).First(&user)
+func (UserService) CheckInfo(userid string) (user *model.User, err error) {
+	model.DB.Where("id = ?", userid).First(&user)
+	return
+}
+
+func (UserService) ChangeInfo(form model.UserChangeRequest) (err error) {
+	res := model.DB.Save(form)
+	err = res.Error
 	return
 }