+++ sdj-be/api_modules/news.md
@@ -2,18 +2,23 @@
 
 ## 学界咨询
 
+Note: module 和 region 永远用int 
+```json
+"module":0|1|2|3|4, // 0:行业资讯，1：学术会议，2：学科竞赛，3：招聘信息,4:全选  仅支持单值传递
+"region":0|1 //domestic|foreign
+```
 #### 获取资讯 `GET {base_url}/api/news`
 - 要求：使用query传递
 
 - Request
     ```json
     {
-        "module":0|1|2|3, // 0:行业资讯，1：学术会议，2：学科竞赛，3：招聘信息, 仅支持单值传递
+        "module":0|1|2|3|4, // 0:行业资讯，1：学术会议，2：学科竞赛，3：招聘信息, 4:全选 仅支持单值传递
         "page":1,
         "limit":20,
         "category":"1100110011", //传入10个学科信息，默认包含全部
-         //可选默认时间倒序
-        "name":"relation|time|popularity", 
+         //默认时间倒序
+        "name":"time|popularity", 
         "order":"increase|reverse"
     }
     ```
@@ -24,15 +29,17 @@
         "data":{
             "category":"110011001",
             "module":0,
+            "total": 100
             "list":[
                 {
                     "id": 1,
                     "title":"题目",
                     "category":"100000000",
                     "module": 0,
-                    "region":"domestic|foreign",
+                    "region":1,
                     "department":"挑战网",
                     "date":"2023-04-03",
+                    "click": 114514
                 }
             ]
         }
@@ -56,7 +63,7 @@
             "title":"题目",
             "category":"100000000",
             "module": 0,
-            "region":"domestic|foreign",
+            "region":1,
             "department":"挑战网",
             "date":"2023-04-03",
             "content":"..." //可能包含富文本
@@ -75,7 +82,7 @@
             "title":"题目",
             "category":"100000000",
             "module": 0,
-            "region":"domestic|foreign",
+            "region":1,
             "department":"挑战网",
             "date":"2023-04-03",
             "content":"..." //可能包含富文本
@@ -103,7 +110,7 @@
             "title":"题目", //可选
             "category":"100000000", //可选
             "module": 0, //可选
-            "region":"domestic|foreign", //可选
+            "region":1, //可选
             "department":"挑战网", //可选
             "date":"2023-04-03", //可选
             "content":"..." //可选
@@ -137,3 +144,42 @@
         }
     }
     ```
+
+#### 搜索 `POST {base_url}/api/news/search`
+- 要求：使用body传递
+- Request
+    ```json
+    {
+        "content": "是学生",//建议前端直接过滤掉非法字符
+        "module":0|1|2|3|4 // 0:行业资讯，1：学术会议，2：学科竞赛，3：招聘信息, 4:全选 仅支持单值传递
+        "page":1,
+        "limit":20,
+        "category":"1100110011", //传入10个学科信息，默认包含全部
+         //默认时间倒序
+        "name":"relativity|time|popularity",   
+        "order":"increase|reverse" //不提供相关性升序
+    }
+  ```
+- Response
+  ```json
+  {
+    "data": {
+        "category": "1111001000",
+        "list": [
+            {
+                "id": 18,
+                "title": "2023年9月全国计算机等级考试报名通知",
+                "module": 3,
+                "department": "挑战网",
+                "click": 0,
+                "date": "2023-06-05",
+                "region": 2,
+                "category": "1010100011"
+            }
+        ],
+        "module": 3,
+        "total": 3
+    },
+    "success": true
+  }
+  ```